sommelier: init at 78.12499.0.0-rc1
sommelier has a lot of dependencies on other Chromium OS packages.  To
manage this mess, I introduced chromiumOSPackages to hold them all,
since most of them won't be useful aside from building other
Chromium OS packages, and chromiumOSPackages.common-mk, which is a
wrapper around stdenv to handle interacting with Chromium OS's
idiosyncratic GN-based build system.

I adapted crosvm's updateScript to become the updateScript for all of
chromiumOSPackages, and pulled crosvm under chromiumOSPackages.  This
means that all Chromium OS packages use approximately the same
versions that are distributed as an upstream release.

There are still a couple of Chromium OS packages in Nixpkgs that
aren't part of this set.  Pulling those in is future work.