start-vm.nix: implement 3-stage init
This uses s6-svscan as stage 2.  Right now all that means is reaping
zombies -- there are no services for it to supervise.  The application
being run in the VM (sommelier hello-wayland in this case) is just run
as a direct child of init.  I felt it didn't belong as a service
because it shouldn't be restarted when it terminates.

I don't use s6-linux-init for a couple of reasons:

- A lot of the reason to use s6-linux-init comes from wanting to do
  clever things to seamlessly redirect logs from the console to the
  file system at some point during stage1.  For VMs, I think we're
  better just logging to the console so logs can be managed in one
  place on the host.

- s6-linux-init is a pain with Nix, because s6-linux-init-maker needs
  to be run as root, and it generates stuff like pipes that can't be
  stored in the Nix store.

coreutils is only used for cp-ing stage3 to the right place.  When
services are implemented, this can be removed and the VMs will have
one less dependency.

To shutdown, we use reboot instead of poweroff because that's what
actually gets crosvm to quit.  reboot(RB_POWER_OFF) seems to be
interpreted by the kernel as halt for some reason I have failed to
determine.  It prints the halt message and everything.
reboot(RB_HALT_SYSTEM) does exactly the same thing.  Maybe it's just
something to do with running in a VM.