chromiumOSPackages.sommelier: fix log spam
Previously, sommelier would output many lines per second saying:

> Failed to send message to virtwl: Success

This made it extremely annoying to use the serial console, and
very difficult to see other log output.

This was happening because Sommelier's check for the return value of
the VIRTWL_IOCTL_SEND ioctl was inverted from what it should be.
That is, the failure message was printed only on success.  If there
actually was an error, no error message would be outputted.

The upstream patch I've added here is a bit vague that this was the
problem, but it does fix the issue.