chromiumOSPackages: improve update.py repo compat
These changes make chromiumOSPackages.updateScript better understand
the Repo manifest format[1].

It now properly distinguishes between paths in the Chromium OS source
tree and paths on the git servers, which are usually not the same.
For example, chromiumos-overlay is located in the source tree at
src/third_party/chromiumos-overlay, but on chromium.googlesource.com
it's located at chromiumos/overlays/chromiumos-overlays.  Components
are now keyed by their location in the Chromium OS source tree (and
packages have all been updated for this change.)

Additionally, it now understands Repo remotes.  This means that it is
now possible to use a Chromium OS component like Minijail, which is
hosted on android.googlesource.com rather than
chromiumos.googlesource.com.  This was not previously possible.

[1]: https://gerrit.googlesource.com/git-repo/+/HEAD/docs/manifest-format.md

Message-Id: <20200530190028.6388-1-hi@alyssa.is>
Reviewed-by: Cole Helbling <cole.e.helbling@outlook.com>