chromiumOSPackages: 81.12871.0.0-rc1 -> 83.13020.0.0-rc1
crosvm now needs the Minijail sources.  However, it'll notice that a
Minijail package is provided via pkg-config, and then not actually use
those sources or build its own Minijail.  It would be nice if the
Minijail we provide could be the same version as in Chromium OS, which
means bringing Minijail under chromiumOSPackages.

A couple of backports are no longer required; one to common-mk and one
to Linux.  A new patch to fix a parallel build failure has been
introduced.

Message-Id: <20200602005619.31128-1-hi@alyssa.is>
Thanks-to: Michael Raskin <7c6f434c@mail.ru>
Reviewed-by: Cole Helbling <cole.e.helbling@outlook.com>
Tested-by: Cole Helbling <cole.e.helbling@outlook.com>