spectrumPackages.rootfs: add s6-rc support
This allows specifying s6-rc services (which can have dependencies,
and be taken up and down) in the same way that s6 services (which are
merely supervised) are declared.

This removes the spectrumcmd mechanism to clean up the boot process --
it could be reintroduced as an s6-rc service if required.

Message-Id: <20210411115740.29615-9-hi@alyssa.is>
Reviewed-by: Cole Helbling <cole.e.helbling@outlook.com>