mdevd: init at 0.1.3.0
By cherry-picking this for Spectrum, we're actually building mdevd in
an unsupported configuration, because it is documented as requiring a
newer version of skalibs than we have.  But it seems to work fine, and
I think we'll be updating soon anyway.

(cherry picked from commit 441f0c894aae2846190724828054e4bad9284579)

Message-Id: <20210411115740.29615-4-hi@alyssa.is>
Reviewed-by: Cole Helbling <cole.e.helbling@outlook.com>