spectrumPackages.rootfs: fix custom path entries
Stage 1 was supposed to be able to take a list of PATH entries to
prepend to the default PATH, but it didn't do anything with them.
Now, it properly adds them to the PATH as early as possible during
boot.

This mechanism isn't currently used anywhere, but is very useful for
debugging.

Message-Id: <20210315135820.6375-1-hi@alyssa.is>
Reviewed-by: Cole Helbling <cole.e.helbling@outlook.com>