spectrumPackages.sys-vms.app: init
This is a very barebones VM.  It's purpose is just to be a
demonstration that other VMs can connect to the router and have their
packets reach the network.  So all it does is infer its IPv4 address,
and the IPv4 address of the router, from the MAC address of its
virtual ethernet device, and configure the network interfaces and
routes appropriately.

vmID is an integer seed we can use to derive things for the VM like IP
and VSOCK addresses.  I don't foresee this sitting around, because I
think it would make more sense for this to be assigned at runtime,
since starting arbitrary VMs at runtime is a goal.  But we'll need
some way to ensure unique addresses.

Message-Id: <20210411115740.29615-16-hi@alyssa.is>
Reviewed-by: Cole Helbling <cole.e.helbling@outlook.com>