spectrumPackages.spectrum-testhost: fix driver bind
The following operations are available to us in /sys/bus/pci/drivers/vfio-pci:
 bind, unbind, new_id, remove_id

bind and unbind control attachment of a device to a driver.
new_id and remove_id control whether a device should be considered
"supported" by a driver.

But when an unassigned device is newly supported by a driver, the
kernel will automatically bind it to that driver.  We were relying on
the new_id operation binding the device to the driver in this way.

But if the driver is in the supported list, but not bound to anything,
this won't happen.  new_id won't do anything, because it's already in
the supported list.  So we can't rely on new_id binding for us, and
need to try bind as well (at least if new_id fails).

To reproduce:

    echo 8086 1502 > /sys/bus/pci/drivers/vfio-pci/new_id
    echo 0000:00:19.0 > /sys/bus/pci/drivers/vfio-pci/unbind

(Use lspci -nn to find the right values for your hardware.)
This will leave you with a device that is supported by vfio-pci but
not bound to it, which would previously cause vm-net to fail to start,
but should not after this change.

Message-Id: <20210421101605.20790-1-hi@alyssa.is>
Reviewed-by: Cole Helbling <cole.e.helbling@outlook.com>