spectrumPackages.makeRootfs: fix messagebus group
/etc/passwd set the messagebus user's primary group to 1, but there
was no group with that id in /etc/group.  The messagebus group had id 4.
The 4 comes from NixOS.  I think it's probably a good idea to use the
same user and group ids where possible -- there's no particular reason
to diverge -- so let's fix it in /etc/passwd instead of changing the
id in /etc/group to 1.

And while we're here, for consistency, let's change the user id of
messagebus to 4 as well, so that's consistent with NixOS as well.
I don't think the user ID is used anywhere else.

This fixes the DBus "Permission denied" errors that have been showing
up in vm-net since forever (but never seemed to cause any real harm...).

Fixes: 52e93c79d82 ("spectrumPackages.rootfs: add dbus configuration")
Message-Id: <20210422010105.415233-1-hi@alyssa.is>
Reviewed-by: Cole Helbling <cole.e.helbling@outlook.com>