cryptsetup: separate binaries from libraries
This reduces closure sizes by making the libraries not depend on the
binaries, which is good for static builds, and (when statically
linked) making the binaries not depend on the libraries, which is good
for static builds.

When static building, we additionally have to disable loading LUKS
plugins from $lib to avoid a reference, which probably wouldn't have
worked anyway.